Modify default seccomp flag to LOG for logging
Changed the default flag from SECCOMP_FILTER_FLAG_SPEC_ALLOW to
SECCOMP_FILTER_FLAG_LOG in order to cause seccomp messages to appear in
the audit log when seccomp is run in trap mode.

This change is required because podman does not currently expose an
interface for passing the SECCOMP_FILTER_FLAG_LOG to crun to change the
default behavior. Rather than upstreaming this commit, it would be
better to modify podman to allow for setting the seccomp flags at
runtime via the --security-opt seccomp json interface and pushing those
changes upstream to the podman project.